collect-wireless-mac: always select first lease from dhcp-server
The script fails at setting the hostname when there are multiple leases
for a mac address.

In line 36 and 37 the hostname gets set from the data the lease. When there
is more than one lease for a specific mac address this fails with the message
"invalid internal item number". More than one lease for a mac address is
possible, if you have more than one SSID on a capsman and a single device
can login into more than one SSID.

Fixes #10

Signed-off-by: Christian Hesse <mail@eworm.de>