global-functions: make notification functions extensible
This allows to add notification functions without overloading functions.
Just add it into the array:

:set ($NotificationFunctions->"fancy-messager") do={
  # notification magic here...
}

Adding functions $SendFancyMessager and/or $SendFancyMessager2 may be
useful. Optionally a function to flush a queue may be required.

A BIG FAT WARNING about function parameters:
Calling a function from array results in $0 for the function name being
skipped. That's why we have to add the function name manually!